data/requirements.txt
selenium>=4.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0